pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fleetintel-mcp"
version = "0.1.0"
description = "Plataforma integrada de inteligência de frota com MCP server, agente LangGraph e WhatsApp"
authors = [
    { name = "Guilherme Taitson", email = "guilhermetaitson@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.11,<3.13"
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "asyncpg>=0.30.0",
    "redis>=5.2.0",
    "supabase>=2.11.0",
    "langgraph>=0.2.70",
    "langchain>=0.3.14",
    "langchain-openai>=0.2.14",
    "langchain-community>=0.3.14",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "tqdm>=4.66.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "black>=24.0.0",
    "mypy>=1.9.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.12.0",
    "ipython>=8.0.0",
    "pre-commit>=3.0.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.28.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E", "F", "I", "N", "W", "UP", "B", "A", "C4", "DTZ", "T10", "ICN", "PIE", "PYI", "PT", "RSE", "RET", "SIM", "ARG", "PTH", "ERA", "PL", "PERF", "RUF"
]
ignore = ["E501"]
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".git",
    "dist",
    "build",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["PLR2004", "S101", "ARG"]

[tool.black]
line-length = 100
target-version = ['py311', 'py312']
skip-string-normalization = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
asyncio_mode = "auto"
markers = [
    "slow",
    "integration",
    "unit",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = "pydantic.mypy"

[tool.hatch.build]
packages = ["mcp_server", "agent", "jobs"]

[tool.hatch.build.targets.wheel]
packages = ["mcp_server", "agent", "jobs"]
